livekit-voice-ai/requirements.txt
livekit-agents[openai,silero,deepgram,cartesia,turn-detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv